Refactor: create a separate class for no-op TraceExporter
